Rechazar usuario ya aceptado terminado

+++ rechazarUsuario.php
@@ -6,11 +6,31 @@
 	// Create connection
 	$mysqli = new mysqli($servername, $username, $password, "album", $dbport);
     $id=$_GET['id'];
-    $sql = mysqli_query($mysqli, "DELETE FROM Usuarios WHERE Usuario='$id'");
+    $sql = mysqli_query($mysqli, "SELECT Email, Nombre, Apellidos, Usuario, Validado FROM Usuarios WHERE Usuario='$id'");
+    if (!$sql) {
+        echo 'No se pudo ejecutar la consulta: ' . mysql_error();
+        exit;
+    }
+	$usuario=mysqli_fetch_row($sql);
+	$sql = mysqli_query($mysqli, "DELETE FROM Usuarios WHERE Usuario='$id'");
     if (!$sql) {
        echo "Error al validar al usuario!";
         exit;
     }else{
-        echo "Usuario validado correctamente!";
+        if($usuario[4]==1){
+            $sql = mysqli_query($mysqli, "DELETE FROM Album WHERE Usuario='$id'");
+            if (!$sql) {
+                echo "Error al validar al usuario!";
+                exit;
+            }
+            $sql = mysqli_query($mysqli, "DELETE FROM Foto WHERE Usuario='$id'");
+            if (!$sql) {
+                echo "Error al validar al usuario!";
+                exit;
+            }
+             echo "Usuario expulsado correctamente!";
+        }else {
+             echo "Usuario validado correctamente!";
+        }
     }
 ?>
